Added mailto in Contact Cards
