--- components/helper/eventsHelper.js
@@ -1,30 +1,0 @@
-import placeholder from "./../../public/images/placeholder.jpg"
-
-
-const events = [
-{
-    image: placeholder,
-    title: "App Development Using Flutter",
-    date: "26th september 2020",
-    speaker: "shashank",
-    discription: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque sit amet justo maximus, tristique lorem sed, facilisis nulla. Nunc ac iaculis massa. Integer ultricies sodales turpis, et convallis dui lacinia sit amet. Duis placerat semper ante ac tempus. In hac habitasse platea dictumst. Vivamus pharetra scelerisque ex at iaculis."
-
-},
-{
-    image: placeholder,
-    title: "App Development Using Flutter",
-    date: "26th september 2020",
-    speaker: "shashank",
-    discription: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque sit amet justo maximus, tristique lorem sed, facilisis nulla. Nunc ac iaculis massa. Integer ultricies sodales turpis, et convallis dui lacinia sit amet. Duis placerat semper ante ac tempus. In hac habitasse platea dictumst. Vivamus pharetra scelerisque ex at iaculis."
-
-},
-{
-    image: placeholder,
-    title: "App Development Using Flutter",
-    date: "26th september 2020",
-    speaker: "shashank",
-    discription: "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Pellentesque sit amet justo maximus, tristique lorem sed, facilisis nulla. Nunc ac iaculis massa. Integer ultricies sodales turpis, et convallis dui lacinia sit amet. Duis placerat semper ante ac tempus. In hac habitasse platea dictumst. Vivamus pharetra scelerisque ex at iaculis."
-
-}]
-
-export default events
